+++ src/components/Layout/LayoutVertical/resourceV2.tsx
@@ -79,6 +79,8 @@ const permissionCollaborator = [
   POLICIES.READ_QUOTATIONPARTNER,
   POLICIES.WRITE_QUOTATIONPARTNER,
   POLICIES.READ_VOUCHERBILLPARTNER,
+  POLICIES.READ_BORROWPRODUCT,
+  POLICIES.READ_STATUSBORROWPRODUCT,
 ];
 
 export const resource: ItemType[] =[ 
@@ -350,7 +352,8 @@ export const resource: ItemType[] =[
     permission: [
       POLICIES.READ_REPORTSALARY, 
       POLICIES.READ_REPORTSALARYPARTNER,
-      POLICIES.READ_VOUCHERSALARYPARTNER
+      POLICIES.READ_VOUCHERSALARYPARTNER,
+      POLICIES.READ_VOUCHERPARTNER,
     ],
     children: [
       {
@@ -476,6 +479,12 @@ export const resource: ItemType[] =[
         path: PATH_APP.vouchers.partner,
         key: PATH_APP.vouchers.partner,
         permission: [POLICIES.READ_VOUCHERBILLPARTNER],//
+      },
+      {
+        label: "Quản lý sản phẩm mượn",
+        path: PATH_APP.product.borrow,
+        key: PATH_APP.product.borrow,
+        permission: [POLICIES.READ_BORROWPRODUCT, POLICIES.READ_STATUSBORROWPRODUCT],//
       },
     ]
   },