+++ src/modules/auth/redux/reducer.ts
@@ -1,6 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { get } from 'lodash';
 import { removeAxiosToken, setAxiosCompanyId, setAxiosToken } from '~/api/requester';
+import { subscribeToken, unSubscribeToken } from '~/modules/user/user.hook';
 const initialState : any = {
     isLoading: false,
 
@@ -29,21 +30,23 @@ export const auth = createSlice({
             state.adapter = null;
         },
         loginSuccess: (state, action: { payload: { token: any; branchId: any,adapter : any } }) => {
-            const { token, branchId,adapter } = action.payload;
+            const { token, branchId, adapter } = action.payload;
             state.token = token;
             state.adapter = adapter;
             setAxiosToken(token);
             setAxiosCompanyId(branchId); // Assuming branchId is the correct property name
             state.isLoading = false;
-            setTimeout(() => {
-                window.location.reload();
-            }, 100);
+            // setTimeout(() => {
+            //     window.location.reload();
+            // }, 100);
+            // token && subscribeToken(token);
           },
         loginFailed: (state, { payload }  : any) => {
             state.loginFailed = payload;
             state.isLoading = false;
         },
         logoutRequest: async() => {
+            unSubscribeToken();
             removeAxiosToken();
             return initialState
         },