+++ src/modules/notification/notification.auth.ts
@@ -1,0 +1,4 @@
+const notificationPermission = {
+
+}
+export default notificationPermission;