+++ src/modules/receiptVoucher/receiptVoucher.hook.ts
@@ -163,7 +163,7 @@ export const useReceiptVoucherQueryParams = () => {
     refCollection = REF_COLLECTION.SUPPLIER
   }
   if(pathname === PATH_APP.vouchers.salaryPartner ){
-    refCollection = compact([REF_COLLECTION.PARTNER, REF_COLLECTION.EMPLOYEE]).join(',');
+    refCollection =  query.get("refCollection")  ||compact([REF_COLLECTION.PARTNER, REF_COLLECTION.EMPLOYEE]).join(',');
     methodType = METHOD_TYPE.VOUCHER_SALARY
   }
   if(pathname === PATH_APP.vouchers.partner ){
@@ -187,7 +187,7 @@ export const useReceiptVoucherQueryParams = () => {
     };
     return [queryParams,onTableChange];
     //eslint-disable-next-line
-  }, [page, limit, keyword, createSuccess, deleteSuccess,startDate, endDate,codeSequence, status, totalAmount, reason,pathname]);
+  }, [page, limit, keyword, createSuccess, deleteSuccess,startDate, endDate,codeSequence, status, totalAmount, reason,pathname,refCollection]);
 };
 
 export const useReceiptVoucherByBillIdQueryParams = (id?: any) => {