+++ src/modules/sale/bill/storeContext/UpdateBillContext.tsx
@@ -46,7 +46,6 @@ export function UpdateBillProvider({
     const [isOpenForm, setIsOpenForm] = useState(false);
     const [isOpenFormPayment, setIsOpenFormPayment] = useState(false);
     // const refCo
-    // console.log(refCollection?.toUpperCase(),'refCollection?.toUpperCase()');
     
   const compareMoney = useMemo(() => pair - totalPrice, [bill]);